Create egg in a whole webpage
Created another webpage to add to the recipes project. I decided to go
with this because it is morning time where I am at right now. It is a
chilly morning and I am enjoying a delicious cup of black coffee from
Wawa.

BLUF: Webpage is created with the boilerplate/foundation.